(Chore): Move model to common
