+++ common/project.json
@@ -1,14 +1,14 @@
 {
-  "name": "models",
+  "name": "common",
   "$schema": "../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
-  "sourceRoot": "models/models",
+  "sourceRoot": "common/common",
   "targets": {
     "lock": {
       "executor": "@nxlv/python:run-commands",
       "options": {
         "command": "poetry lock --no-update",
-        "cwd": "models"
+        "cwd": "common"
       }
     },
     "add": {
@@ -27,7 +27,7 @@
       "executor": "@nxlv/python:build",
       "outputs": ["{projectRoot}/dist"],
       "options": {
-        "outputPath": "models/dist",
+        "outputPath": "common/dist",
         "publish": false,
         "lockedVersions": true,
         "bundleLocalDependencies": true
@@ -46,21 +46,21 @@
     },
     "lint": {
       "executor": "@nxlv/python:flake8",
-      "outputs": ["{workspaceRoot}/reports/models/pylint.txt"],
+      "outputs": ["{workspaceRoot}/reports/common/pylint.txt"],
       "options": {
-        "outputFile": "reports/models/pylint.txt"
+        "outputFile": "reports/common/pylint.txt"
       },
       "cache": true
     },
     "test": {
       "executor": "@nxlv/python:run-commands",
       "outputs": [
-        "{workspaceRoot}/reports/models/unittests",
-        "{workspaceRoot}/coverage/models"
+        "{workspaceRoot}/reports/common/unittests",
+        "{workspaceRoot}/coverage/common"
       ],
       "options": {
         "command": "poetry run pytest tests/",
-        "cwd": "models"
+        "cwd": "common"
       },
       "cache": true
     }
